Разбила код на модули и раскидала в соотвутствующие файлы
